web: Use fieldsets on new proposal page
This is to more clearly separate the new proposal, continuation
request, and copy options.  Continuation request moved above copy.
Also the formnovalidate attribute was removed from the continuation
request and copy submit buttons so that the browser can indicate
that affiliation is required.  The required attribute was removed
from the title field.