web: Check role relvant in case edit view methods
This adds a new decorator 'with_relevant_text_role' which can
be used to check that a role is relevant to the current proposal.
I.e. if the text role class specifies a proposal section for the
role, then that section must appear in the section order for
the proposal type of the current proposal.