web: Add call review clash search
This applies a clash tool-like search between all of the proposals
involved in a review process.  A second page gives a detailed comparison
of two proposals, showing which targets have similar positions.

Note that the search methods have been placed in the clash tool in order
to make use of some of its information, but if it is necessary to
perform this search in a facility which does not use the clash tool,
then the functionality could be moved elsewhere.