web: Apply page size limit to uploaded PDF files
This adds maximum major and minor dimensions to the configuration
file along with a text description to be shown in the web interface.
The template configuration size is slightly larger than A4 or Letter
size in each dimension, so pages of size A4 or Letter should be
accepted.

The maximum page count and size are shown in a list on the PDF upload
page.  (This replaces showing the maximum page count in the error
message if exceeded.)