web: Use inst. abbreviations in selection lists
This updates the hedwig.view.query.truncate function to accept
an abbreviation argument.  In addition it reduces the length
by one when truncated (to allow space for the ellipsis) and
trims the string (in case it was truncated at a space).
A test for this function has been added.

The JSON query data now includes attributes "text_full"
and "text_abbr" containing the unabbreviated/untruncated
text and any given abbreviations.  The Selectize controls
use these attributes for their search function,
in order to allow matching with information not shown
in the abbreviated/truncated form, or an abbreviation
if the text is shown in full.