email: Avoid format=flowed when using QP content
When the content transfer encoding is quoted printable, avoid using
format=flowed.  Unfortunately the "render_email_template" function
(from hedwig.email.format) applies "wrap_email_text" to prepare the
message for sending with format=flowed.  In order to allow both
schemes to be used and to maintain compatibility with existing
entries in the message table, apply a new function "unwrap_email_text"
when preparing a message to be sent as quoted printable.