email: Improve spacing in _unwrap_email_template
Attempt to improve the logic for when a space should be inserted
on unwrapping lines.  Does this by recognizing lines containing
just a single '{% ... %}' statement and assuming it does not
generate output.  When such statements are if-elif-else-endif
blocks, record the initial status to be able to restore it
for each branch.