web: Show all submitted states on reviewers page
This is so that we retain the ability to send a "thank you"
message to invited reviewers even if the review process has
been completed.  Now include a "can_edit" attribute in the
reviewer records prepared for the HTML template to indicate
which reviews are editable given the proposal's state.
Also filter the reviews for which notification is suggested
by state so that we do not suggest notifying reviewers after
the process is completed.