web: Show links to reviews from prev. prop. list
For each entry in the preview proposal list, check authorization
to view the proposal and its reviews (using the new auth functions
via PrevProposal objects).  Use this information to determine
whether to show the proposal code as a link, and whether to include
a link to the proposal reviews page.  This is to make it easier
to, for example, see the previous reviews of a continuation
request.

Note: we assume here that the proposal membership of the previous
proposals will be similar to that of the proposal being viewed,
and so use its members list for the authorization functions.
In any case the permission to actually view the proposal or
reviews will be checked again when the corresponding page
is loaded.