web: Apply call aff/n overrides at proposal stage
Update "search_affiliation" database queries in the proposal
part of the generic facility view class to supply the call ID
via the "with_weight_call_id" parameter.  This is to allow
the affiliation types (and hidden parameter) to be overridden
on a call-specific basis while the call for proposals is still
open.  (Previously these overrides were only applied during
the review process.)