web: Add formatting for retrieved pub. titles
Sometimes publication services return titles including HTML-like
formatting (SUP, SUB, ASTROBJ) or LaTeX formatting (e.g. ^, _).
Therefore add a template filter to attempt to process this and apply
it to titles in the "previous proposals and publications" list.
(I.e. this is currently only applied to retrieved titles rather than
descriptions entered directly.)

The formatting routine applies a series of simple regular expression
splits to match pieces of markup, recursing on the non-matching
sections.  (So nesting is not supported.)  Formatting is applied
in a try-except block, returning the unformatted text in the event
of an error and logging the problematic text.