web: Trap calculator errors from query strings
Now record whether the source of "fetched_input" is a query string
in a calculation link.  If so, trap exceptions from performing the
calculation and formatting the inputs.  This is because the query
string may have originated outside the Hedwig system, so we should
avoid logging an error in this case.