web: Allow sort/converting proposal target coords
This adds the ability to sort the entries of the target objects table
of a proposal using the existing "enable_table_sorting" function
from "hedwig_util.js".  This function has been extended to add support
for a button to restore the original ordering.

Additionally buttons were added to convert the coordinates between
the available systems (currently ICRS and Galactic).  The information
for this is pre-prepared by calling TargetCollection.to_formatted_collection
with a new parameter "with_extra".  This in turn calls a new function
"format_coord_all_systems" in the hedwig.astro.coord module which
prepares a dictionary with coordinates (formatted and as degrees)
in all available systems.  The format_coord function gained a
"fixed_precision" parameter which is used here to prevent excess
precision in the formatted representation of converted values.