publication: Add function to look up arXiv via ADS
While this function does work, the wrong year may be returned.
(The year of final publication rather than upload to arXiv.)

This function has been added in case it ever becomes necessary
to look up arXiv IDs via ADS, but for now it is not used -- the
"get_pub_info_arxiv" function is used instead.