type: Add "thank" attribute to BaseReviewerRole
This is to control whether or not the web interface should offer
to send a "thank you" message to reviewers.  Previously it was
assumed that this should be when the reviewer was of an invited
role.  Now as a separate setting it can be applied to peer reviewers
as well as (invited) external reviewers.