web: Add feature tuple to control observer flag
This adds a new tuple "FacilityFeatures" to control general facility
feature settings.  The tuple is accessed by a new facility class
method "get_features".

At the moment it just contains an attribute "member_observer",
indicating whether the proposal member list should include the
"observer" option for members.  If false then this option is not
shown in the web interface, and is not retained when a proposal
is copied.