type: Add GroupType perm. attribute "hidden_call"
This adds a new permission attribute "hidden_call" to the information
for values of the GroupType enum type.  This attribute is enabled for
the HIDDEN_CALL type.  This allows the "auth" module to handle permission
to view hidden calls in the normal manner rather than having to specifically
recognize the HIDDEN_CALL type.  It also allows the group view page to
include this information in the list of the group's permissions.

(This page now uses a counter object to automatically count the
permissions and this change was also applied to the site group view page.)