web: Update person profile view page
Various minor updates to the profile page.  Change the title
to "Your Profile" (if for the current user) rather than including
a note on the page.  Add an indication of administrative access
(if viewed as an administrator).  Make the institution name
a link (rather than having a separate view link below).
Show grid tables for group membership to make it easier
to compare the person's membership of different groups.