web: Add pages allowing affiliation/group copying
These new pages allow list of affiliations and group membership to
be copied from one queue to another.  This uses a new utility
function "compare_collections" (and a new helper method
ResultCollection.get_value) in order to prepare a list of changes
for display.  The updated collections are then saved to the database
using the usual sync methods.

This is to allow for the usage where multiple queues are created to
allow for various extra calls for proposals, but they have similar
settings.  A possible future enhancement would be to allow the user
to select which of the changes they wish to apply.