jcmt: Allow USB/LSB only receivers in Het. ITC
Assume that USB is available if we have LO-based Trx data (t_rx_lo
is true) and USB Trx data are present (t_rx_usb is not none), and
likewise for LSB.  Then if we are offering an IF option
(i.e. when t_rx_lo is true) enable only the relevant options
in the sideband selection control.  (The "best sideband" option
is currently always available so that it can stay selected when
changing receivers.)