web: Add template filter for regional indicators
This adds a template filter to convert a 2-character country code
to the corresponding Unicode regional indicator symbol.  By default
this returns an <abbr> tag including the country name in the title.

A new "compat" function "char" was added.  This is chr() in Python 3
and unichr() in Python 2 to ensure unicode values are returned
and extended code points are supported.

(Currently not used anywhere, but could potentially be useful
wherever full institution descriptions are too long.)