type: Add recipient_public to MessageThreadType
This new attribute can be used to indicate when messages of a given
thread type should have all recipients listed in the "To" header.
(Previously only those with public addresses, or a single recipient,
were listed).  Apply this to the PROPOSAL_REVIEW thread type so
that everyone is included for administrative messages regarding
proposal review (for immediate review proposals).