jcmt: Exclude peer rating if cttee reviews present
Previously calls were either peer-reviewed (i.e. Rapid Turnaround) or
committee-reviewed (with external reviewers, e.g. Regular calls).
External review ratings were considered informational and not included
in the overall rating of a proposal.  To apply the same logic to calls
where peer reviews have been used in place of external reviews, we need
to check which types of reviews are present when calculating the overall
rating.  Unfortunately there is currently no more structured way to
determine whether the review coordinator will have decided to use peer
or external reviews for a particular call.  In order to allow this
determination to be made, the complete set of reviewers can now be
passed to the "get_review_rating_weight_function" method.  The JCMT
version of this method checks whether any reviews with the
"jcmt_expertise" flag are present (i.e. committee reviews), and if so,
it will not provide a weighting for reviews with the
"jcmt_peer_expertise" flag.